requirements.txt
jsonpickle
psutil
requests
tqdm
elasticsearch
elasticsearch_dsl
SimConnect==0.4.23
pytest
pandas
mss
Pillow
opencv-python
Flask
Flask-RESTful
flask-cors
python-slugify
moviepy
tensorflow
numpy